ISIS-2726: collapses SudoServiceDefault into SudoServiec
No need to support multiple implementations here
