+++ api/applib/src/main/java/org/apache/isis/applib/services/sudo/SudoService.java
@@ -18,15 +18,29 @@
  */
 package org.apache.isis.applib.services.sudo;
 
+import java.util.List;
 import java.util.concurrent.Callable;
 import java.util.function.UnaryOperator;
 
+import javax.inject.Inject;
+import javax.inject.Named;
+
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.context.annotation.Primary;
+import org.springframework.core.annotation.Order;
+import org.springframework.stereotype.Service;
+
+import org.apache.isis.applib.annotation.OrderPrecedence;
 import org.apache.isis.applib.services.iactnlayer.InteractionContext;
+import org.apache.isis.applib.services.iactnlayer.InteractionService;
+import org.apache.isis.applib.services.iactnlayer.InteractionTracker;
+import org.apache.isis.applib.services.iactnlayer.ThrowingRunnable;
 import org.apache.isis.applib.services.user.RoleMemento;
 import org.apache.isis.applib.services.user.UserService;
-import org.apache.isis.applib.services.iactnlayer.ThrowingRunnable;
 
 import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+import lombok.val;
 
 /**
  * Allows a block of code to be executed within an arbitrary
@@ -35,7 +49,7 @@ import lombok.NonNull;
  *
  * <p>
  * Most typically this service is used to temporarily change the
- * &qout;who&quot;, that is the user reported by the {@link UserService}'s
+ * &quot;who&quot;, that is the user reported by the {@link UserService}'s
  * {@link UserService#currentUser() getUser()} - hence the name SudoService.
  * But the user's locale and timezome can also be changed, as well as the time
  * reported by {@link org.apache.isis.applib.services.clock.ClockService}.
@@ -48,27 +62,26 @@ import lombok.NonNull;
  *
  * @since 1.x revised for 2.0 {@index}
  */
-public interface SudoService {
+@Service
+@Named("isis.applib.SudoService")
+@Order(OrderPrecedence.MIDPOINT)
+@Primary
+@Qualifier("Default")
+@RequiredArgsConstructor(onConstructor_ = {@Inject})
+public class SudoService {
 
     /**
      * If included in the list of roles, then will disable security checks (can view and use all object members).
      */
-    RoleMemento ACCESS_ALL_ROLE =
+    public static RoleMemento ACCESS_ALL_ROLE =
             new RoleMemento(
                     SudoService.class.getName() + "#accessAll",
                     "Sudo, can view and use all object members.");
 
 
-    /**
-     * Executes the supplied {@link Callable} block, within the provided
-     * {@link InteractionContext}.
-     *
-     * @param sudoMapper - maps the current {@link InteractionContext} to the sudo one
-     * @since 2.0
-     */
-    <T> T call(
-            final @NonNull UnaryOperator<InteractionContext> sudoMapper,
-            final @NonNull Callable<T> supplier);
+    private final InteractionService interactionService;
+    private final InteractionTracker interactionTracker;
+    private final List<SudoServiceListener> sudoListeners;
 
     /**
      * Executes the supplied {@link Callable} block, within the provided
@@ -77,11 +90,39 @@ public interface SudoService {
      * @param sudoMapper - maps the current {@link InteractionContext} to the sudo one
      * @since 2.0
      */
-    default void run(
+    public <T> T call(
+            final @NonNull UnaryOperator<InteractionContext> sudoMapper,
+            final @NonNull Callable<T> callable) {
+
+        val currentInteractionLayer = interactionTracker.currentInteractionLayerElseFail();
+        val currentInteractionContext = currentInteractionLayer.getInteractionContext();
+        val sudoInteractionContext = sudoMapper.apply(currentInteractionContext);
+
+        try {
+            for (val sudoListener : sudoListeners) {
+                sudoListener.beforeCall(currentInteractionContext, sudoInteractionContext);
+            }
+
+            return interactionService.call(sudoInteractionContext, callable);
+
+        } finally {
+            for (val sudoListener : sudoListeners) {
+                sudoListener.afterCall(sudoInteractionContext, currentInteractionContext);
+            }
+        }
+    }
+
+    /**
+     * Executes the supplied {@link Callable} block, within the provided
+     * {@link InteractionContext}.
+     *
+     * @param sudoMapper - maps the current {@link InteractionContext} to the sudo one
+     * @since 2.0
+     */
+    public void run(
             final @NonNull UnaryOperator<InteractionContext> sudoMapper,
             final @NonNull ThrowingRunnable runnable) {
         call(sudoMapper, ThrowingRunnable.toCallable(runnable));
     }
-
 
 }
